Separate portable-only portion of RegisteredWaitHandle
